Alert user about expired GH credentials
Moving the known error codes from a backend specific location to the 
error package so it can be used for both Tauri and forges.

The other changes just make error handling behave more as intended.
